requirements.txt
# AI Gym & Fitness Assistant — Python dependencies
# install with: pip install -r requirements.txt

# web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# data validation
pydantic==2.7.1

# AI/ML (core)
numpy==1.26.4
scikit-learn==1.4.2
opencv-python==4.9.0.80
mediapipe==0.10.14

# deep learning (optional — heavy, install only if needed)
# torch==2.3.0
# tensorflow==2.16.1

# NLP
transformers==4.40.1
# huggingface-hub==0.22.2

# database
pymongo==4.7.1
motor==3.4.0         # async MongoDB driver

# IoT communication
paho-mqtt==2.0.0

# HTTP client (for external APIs)
httpx==0.27.0

# environment variables
python-dotenv==1.0.1

# utilities
python-multipart==0.0.9   # for file uploads
